Removing action labels from transitions.
PRISM (and by extension, PARAM) does not support different
synchronization directives for the same guard. As we use only the
current state as guard for the commands, this means the synchronization
directives should be associated to the states, and not the transitions.

Furthermore, the generated models are temporary and self-contained, so they
cannot be used as documentation and need not be used for synchronizing
modules.