Setting up an FDTMC initial state to be the first one created.
There was some discussion whereby four approaches where considered:

  - infer the initial state from the FDTMC's topology (the initial state
    would be the one with no incoming transitions except for self-loops);
  - consider a state labeled "init" as the initial one;
  - take the firstly created state to be the initial one; or
  - expose a setInitialState() method.

The first would be unnecessarily expensive, as the information on which
state is initial is available at FDTMC creation time (during conversion
from UML behavioral models). Nonetheless, it is the only one which
fully respects the DTMC properties.

The fourth would be error prone, since it would be possible to build an
FDTMC with no initial state.

The second would be semantically equivalent to the fourth, but without
type safety (the "initial" property would be specified as a string).

Finally, the third is simple and ensures the existence of an initial state.
The developer only needs to create the transitions and states starting
from the initial state of the chain. This seems reasonable, since the
FDTMCs are to be generated from top-to-bottom reading of UML behavioral
models.