Getting rid of classes and tests which became obsolete.
The new method for reliability evaluation using ADD arithmetics was
incompatible with most of the previous implementation.

Also, tests were "promoted" to the same level as "src", so that their
folder can be directly added to the build path. This is so that tests
can belong to the same package as the classes being tested, in order to
leverage package-private visibility.