Using presence conditions' equivalence classes to encode variability.
We were previously using indices of RDG nodes themselves, which can lead
to a greater number of variables than necessary in case the behavior of
a given feature is scattered in more than one RDG node.

Thus we turned to an approach of encoding based on presence conditions.
Presence conditions are not directly used, since they can have
logical operators (leading to invalid PARAM identifiers). What we do is
generate indices for presence conditions (similar to what we do with RDG
nodes). As a given index is consistently attributed to equivalent
conditions, this gives rise to indices for equivalence classes.

As of now, the equivalence relation is merely string equality, but
logical equivalence, simultaneous presence and other richer relations
could also be used in the future to further reduce the amount of
variables.