Added option for suppressing reliability report.
The flag --suppress-report causes the tool to evaluate reliabilities
(using whatever strategy is selected) but not print them to stdout.

The practical effect is that the total running time becomes entirely
due to analysis tasks. Family-based and feature-family-based analyses
do not need to enumerate the configurations at all, and the others do it
just for the sake of product-based evaluation.